+++ small_glim/include/small_glim/preprocess/cloud_preprocessor.hpp
@@ -57,7 +57,7 @@ public:
     PreprocessedFrame::Ptr preprocess(const RawPoints::ConstPtr raw_points);
 
 private:
-    std::vector<int> find_neighbors(const Eigen::Vector4d* points, const int num_points, const int k) const;
+    std::vector<size_t> find_neighbors(const Eigen::Vector4d* points, const size_t num_points, const size_t k) const;
 
 private:
     std::unique_ptr<CloudPreprocessorParams> params;
