+++ small_glim/src/preprocess/time_keeper.cpp
@@ -133,8 +133,8 @@ void TimeKeeper::replace_points_stamp(const small_glim::RawPoints::Ptr points) {
         points->times.resize(points->size(), 0.0);
         const double scan_duration = estimate_scan_duration(points->stamp);
         if (scan_duration > 0.0) {
-            for (int i = 0; i < points->size(); i++) {
-                points->times[i] = scan_duration * static_cast<double>(i) / points->size();
+            for (size_t i = 0; i < points->size(); i++) {
+                points->times[i] = scan_duration * static_cast<double>(i) / static_cast<double>(points->size());
             }
         }
